Create even_or_odd_num1.cpp
Theory :
This C++ program checks whether a given integer is even or odd. 
It prompts the user to enter a number, which is stored in the variable num. 
The program then uses the modulo operator (%) to determine the remainder when num is divided by 2. 
If the remainder is 0 (num % 2 == 0), the number is even; otherwise, it is odd. The result is printed accordingly using cout. 
This simple logic demonstrates the use of conditionals (if-else) and modular arithmetic.